libdsd2pcm cleaned up
